frontend: move all module css into styles folder
